Add functional reviews domain

+++ app/Http/Controllers/BookController.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreBookRequest;
-use Illuminate\Http\Request;
+use Illuminate\Http\Exceptions\HttpResponseException;
 use App\Models\Book;
 use App\Http\Resources\BookResource;
+use Illuminate\Support\Facades\Log;
+
 
 class BookController extends Controller
 {
@@ -17,6 +19,9 @@ class BookController extends Controller
     public function store(StoreBookRequest $request)
     {
         $book = Book::create($request->validated());
+        info('Hellow world info');
+        Log::info('This is an info log');
+
 
         $books = Book::all();
         return BookResource::collection($books);
@@ -32,6 +37,11 @@ class BookController extends Controller
 
     public function destroy(Book $book)
     {
+        if ($book->reviews()->exists()) {
+            throw new HttpResponseException(response()->json([
+                'message' => 'Dit boek kan niet worden verwijderd omdat er nog reviews aan gekoppeld zijn.'
+            ], 422));
+        }
         $book->delete();
         return response()->json(['message' => 'Boek succesvol verwijderd']);
     }
